--- src/Microsoft.AspNetCore.HttpsPolicy/HttpsPolicyServiceCollectionExtensions.cs
@@ -1,17 +1,0 @@
-﻿// Copyright (c) .NET Foundation. All rights reserved.
-// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
-
-using System;
-using Microsoft.AspNetCore.HttpsPolicy;
-using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.DependencyInjection.Extensions;
-
-namespace Microsoft.AspNetCore.Builder
-{
-    /// <summary>
-    /// Extension methods for the HttpsPolicy middleware.
-    /// </summary>
-    public static class HttpsPolicyServiceCollectionExtensions
-    {
-    }
-}
